added a real home page
